refactor e2e tests a bit
