+++ 3rdparty/python/constraints.txt
@@ -1,17 +1,21 @@
-# Generated by ./build-support/manage_virtualenv.sh on Sun 25 Jul 2021 11:04:37 PM IST
+# Generated by ./build-support/manage_virtualenv.sh on Sun 25 Jul 2021 11:30:20 PM IST
 # DO NOT EDIT BY HAND
+ansicolors==1.1.8
 Arpeggio==1.10.2
 attrs==20.3.0
 blessed==1.17.6
 boto3==1.18.6
 botocore==1.21.6
+bugout==0.1.14
 certifi==2021.5.30
 chalice==1.22.4
 charset-normalizer==2.0.3
 click==7.1.2
 confluent-kafka==1.7.0
 dill==0.3.4
+fasteners==0.16
 grpcio==1.39.0
+humbug==0.1.9
 hypothesis==6.14.3
 idna==3.2
 importlib-metadata==4.6.1
@@ -30,17 +34,21 @@ mypy-boto3-sns==1.18.6
 mypy-boto3-sqs==1.18.6
 mypy-boto3-ssm==1.18.6
 mypy-extensions==0.4.3
-packaging==21.0
+packaging==20.9
+pantsbuild.pants==2.6.0rc1
 parver==0.3.1
+pex==2.1.42
 pip==21.1.3
 pluggy==0.13.1
 protobuf==3.17.3
+psutil==5.8.0
 pulumi==3.8.0
 pulumi-aws==4.6.0
 pulumi-docker==3.0.0
 pulumi-kafka==3.0.1
 pulumi-random==4.2.0
 py==1.10.0
+pydantic==1.8.2
 pydgraph==21.3.0
 PyJWT==2.1.0
 pyparsing==2.4.7
@@ -53,11 +61,16 @@ redis==3.5.3
 requests==2.26.0
 s3transfer==0.5.0
 semver==2.13.0
-setuptools==47.1.0
+setproctitle==1.2.2
+setuptools==56.2.0
 six==1.16.0
 sortedcontainers==2.4.0
 toml==0.10.2
-typing-extensions==3.10.0.0
+types-PyYAML==5.4.3
+types-requests==2.25.0
+types-setuptools==57.0.0
+types-toml==0.1.3
+typing-extensions==3.7.4.3
 urllib3==1.26.6
 wcwidth==0.2.5
 wheel==0.36.2
